docs(): 整理 brick-result general-tooltip brick-humanize-time 构件文档 refs BRICK_STORE-2911
